+++ core/src/components/cat-datepicker/cat-datepicker.tsx
@@ -1,4 +1,4 @@
-import { Component, Event, Method, EventEmitter, Prop, h } from '@stencil/core';
+import { Component, Event, EventEmitter, Method, Prop, Watch, h } from '@stencil/core';
 import flatpickr from 'flatpickr';
 import weekSelectPlugin from 'flatpickr/dist/plugins/weekSelect/weekSelect';
 import { ErrorMap } from '../cat-form-hint/cat-form-hint';
@@ -165,38 +165,25 @@ export class CatDatepickerFlat {
    */
   @Event() catBlur!: EventEmitter<FocusEvent>;
 
+  @Watch('value')
+  onValueChanged(value: string) {
+    if (value) {
+      this.pickr?.setDate(value, false);
+      this.catChange.emit(value);
+    } else {
+      this.pickr?.clear(false);
+      this.catChange.emit(undefined);
+    }
+  }
+
+  @Watch('disabled')
+  onDisabledChaned(value: boolean) {
+    this.pickr?.set('clickOpens', !value);
+  }
+
   componentDidLoad() {
-    const input = this.input;
-    if (input) {
-      const locale = getLocale(i18n.getLocale());
-      const format = getFormat(i18n.getLocale(), this.mode);
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      const plugins = this.mode === 'week' ? [new (weekSelectPlugin as any)({})] : [];
-      this.pickr = flatpickr(input, {
-        locale,
-        plugins,
-        altInput: true,
-        prevArrow: '←',
-        nextArrow: '→',
-        dateFormat: this.dateFormat,
-        altFormat: format,
-        ariaDateFormat: format,
-        mode: this.mode === 'daterange' ? 'range' : 'single',
-        minDate: this.min,
-        maxDate: this.max,
-        enableTime: this.mode === 'time' || this.mode === 'datetime',
-        noCalendar: this.mode === 'time',
-        weekNumbers: true,
-        minuteIncrement: this.step,
-        onChange: (dates, dateStr, flatpickr) => {
-          if (this.mode === 'week') {
-            this.value = dates[0] ? flatpickr.config.getWeek(dates[0]).toString() : undefined;
-          } else {
-            this.value = dateStr;
-          }
-          this.catChange.emit(this.value);
-        }
-      });
+    if (this.input) {
+      this.pickr = this.initDatepicker(this.input);
     }
   }
 
@@ -248,11 +235,7 @@ export class CatDatepickerFlat {
         nativeAttributes={this.nativeAttributes}
         onCatChange={e => {
           e.stopPropagation();
-          this.pickr?.setDate(e.detail);
-          if (this.value !== (e.detail || undefined)) {
-            this.value = e.detail || undefined;
-            this.catChange.emit(this.value);
-          }
+          this.value = e.detail || undefined;
         }}
         onCatFocus={e => {
           e.stopPropagation();
@@ -264,6 +247,38 @@ export class CatDatepickerFlat {
         }}
       ></cat-input>
     );
+  }
+
+  private initDatepicker(input: HTMLInputElement): flatpickr.Instance {
+    const locale = getLocale(i18n.getLocale());
+    const format = getFormat(i18n.getLocale(), this.mode);
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    const plugins = this.mode === 'week' ? [new (weekSelectPlugin as any)({})] : [];
+    return flatpickr(input, {
+      locale,
+      plugins,
+      altInput: true,
+      prevArrow: '←',
+      nextArrow: '→',
+      dateFormat: this.dateFormat,
+      altFormat: format,
+      ariaDateFormat: format,
+      mode: this.mode === 'daterange' ? 'range' : 'single',
+      minDate: this.min,
+      maxDate: this.max,
+      enableTime: this.mode === 'time' || this.mode === 'datetime',
+      noCalendar: this.mode === 'time',
+      weekNumbers: true,
+      minuteIncrement: this.step,
+      clickOpens: !this.disabled,
+      onChange: (dates, dateStr, flatpickr) => {
+        if (this.mode === 'week') {
+          this.value = dates[0] ? flatpickr.config.getWeek(dates[0]).toString() : undefined;
+        } else {
+          this.value = dateStr || undefined;
+        }
+      }
+    });
   }
 
   private get dateFormat(): string {