+++ core/src/components/cat-radio-group/cat-radio-group.tsx
@@ -105,8 +105,10 @@ export class CatRadioGroup {
   @Listen('input')
   onInput(event: MouseEvent): void {
     const radio = this.catRadioGroup.find(radio => radio === event.target);
-    this.value = radio?.checked ? radio?.value : undefined;
-    this.catChange.emit(this.value);
+    if (radio?.localName === 'cat-radio') {
+      this.value = radio?.checked ? radio?.value : undefined;
+      this.catChange.emit(this.value);
+    }
   }
 
   @Listen('focus', { capture: true })