+++ core/src/components/cat-tabs/cat-tabs.tsx
@@ -32,7 +32,7 @@ export class CatTabs {
   @Prop() tabsAlign: 'left' | 'center' | 'right' | 'justify' = 'left';
 
   @Watch('activeTabId')
-  onActiveTabChanged(newActiveTab: string): void {
+  onActiveTabIdChanged(newActiveTab: string): void {
     const activeTab = this.tabs.find(value => value.id === newActiveTab);
     activeTab?.click();
   }