+++ tokens/config.js
@@ -1,4 +1,4 @@
-const StyleDictionary = require('style-dictionary');
+const StyleDictionary = require('style-dictionary-utils');
 const tinycolor = require("tinycolor2");
 
 StyleDictionary.registerFileHeader({
@@ -6,17 +6,42 @@ StyleDictionary.registerFileHeader({
   fileHeader: () => ['Auto-generated file. Do not edit directly.']
 });
 
-StyleDictionary.registerTransform({
-  type: 'value',
-  name: 'cat/size',
-  matcher: token => token.attributes.category === 'size',
-  transformer: token => `${token.value}${token.attributes.unit || 'rem'}`
+StyleDictionary.registerFormat({
+  name: 'json/designTokens',
+  formatter: function ({ dictionary }) {
+    const set = (obj, path, token) => {
+      const [key, ...restPath] = path;
+      obj[key] = restPath.length ? set(obj[key] || {}, restPath, token) : {
+        $type: token.$type,
+        $value: token.value
+      };
+      return obj;
+    }
+    const tokens = dictionary.allTokens.reduce((acc, token) =>
+      set(acc, token.path, token)
+    , {});
+    return JSON.stringify(tokens, null, 2);
+  }
+});
+
+StyleDictionary.registerFormat({
+  name: 'json/cssProp',
+  formatter: function ({ dictionary }) {
+    const tokens = dictionary.allTokens.reduce((acc, token) => {
+      acc[token.cssProp] = {
+        $type: token.$type,
+        $value: token.value
+      };
+      return acc;
+    }, {});
+    return JSON.stringify(tokens, null, 2);
+  }
 });
 
 StyleDictionary.registerTransform({
   type: 'value',
   name: 'cat/rgbParts',
-  matcher: token => token.attributes.category === 'color',
+  matcher: token => token.$type === 'color',
   transformer: token => {
     var rgb = tinycolor(token.value).toRgb();
     return `${rgb.r}, ${rgb.g}, ${rgb.b}`;
@@ -27,15 +52,23 @@ StyleDictionary.registerTransform({
   type: 'value',
   name: 'cat/cssProp',
   transitive: true,
-  matcher: token => !!token.attributes.cssProp,
-  transformer: token => `var(--cat-${token.attributes.cssProp}, ${token.value})`
+  matcher: token => !!token.cssProp,
+  transformer: token => `var(--cat-${token.cssProp}, ${token.value})`
+});
+
+StyleDictionary.registerTransform({
+  type: 'value',
+  name: 'cat/jsNumber',
+  transitive: true,
+  matcher: token => token.$type === 'dimension' || token.$type === 'duration',
+  transformer: token => parseFloat(token.value)
 });
 
 module.exports = {
-  source: ['src/**/*.js'],
+  source: ['src/**/*.json'],
   platforms: {
     js: {
-      transforms: ['attribute/cti', 'name/cti/camel', 'size/rem', 'color/hex'],
+      transforms: ['attribute/cti', 'name/cti/camel', 'color/hex', 'cat/jsNumber'],
       buildPath: 'dist/js/',
       files: [{
         destination: 'variables.js',
@@ -52,7 +85,7 @@ module.exports = {
       }]
     },
     css: {
-      transforms: ['attribute/cti', 'name/cti/kebab', 'time/seconds', 'content/icon', 'color/css', 'cat/size', 'cat/rgbParts'],
+      transforms: ['attribute/cti', 'name/cti/kebab', 'content/icon', 'color/css', 'cat/rgbParts'],
       prefix: 'cat',
       buildPath: 'dist/css/',
       files: [{
@@ -65,7 +98,7 @@ module.exports = {
       }]
     },
     scss: {
-      transforms: ['attribute/cti', 'name/cti/kebab', 'time/seconds', 'content/icon', 'color/css', 'cat/size', 'cat/rgbParts', 'cat/cssProp'],
+      transforms: ['attribute/cti', 'name/cti/kebab', 'content/icon', 'color/css', 'cat/rgbParts', 'cat/cssProp'],
       prefix: 'cat',
       buildPath: 'dist/scss/',
       files: [{
@@ -79,11 +112,38 @@ module.exports = {
       }]
     },
     json: {
+      transforms: ['name/cti/kebab'],
       buildPath: 'dist/json/',
       files: [{
         destination: 'variables.json',
         format: 'json/nested'
       }]
     },
+    zeroheight: {
+      transforms: ['name/cti/kebab'],
+      buildPath: 'dist/export/',
+      files: [{
+        destination: 'zeroheight.json',
+        format: 'json/designTokens'
+      }]
+    },
+    figma: {
+      transforms: ['name/cti/kebab', 'dimension/pixelUnitless'],
+      buildPath: 'dist/export/',
+      files: [{
+        destination: 'figma.json',
+        format: 'json/designTokens',
+        filter: (token) => token.$type !== 'asset'
+      }]
+    },
+    theme: {
+      transforms: ['name/cti/kebab'],
+      buildPath: 'dist/export/',
+      files: [{
+        destination: 'theme.json',
+        format: 'json/cssProp',
+        filter: (token) => token.hasOwnProperty('cssProp')
+      }]
+    }
   }
 };