+++ tokens/src/opacity.js
@@ -1,7 +1,0 @@
-module.exports = {
-  opacity: {
-    disabled: { value: 0.65, attributes: { cssProp: "opacity-disabled" } },
-    backdrop: { value: 0.4, attributes: { cssProp: "opacity-backdrop" } },
-    tooltip: { value: 1, attributes: { cssProp: "opacity-tooltip" } }
-  }
-}