+++ tokens/src/size/modal.js
@@ -1,9 +1,0 @@
-module.exports = {
-  size: {
-    modal: {
-      s: { value: 400, attributes: { unit: "px" } },
-      m: { value: 600, attributes: { unit: "px" } },
-      l: { value: 800, attributes: { unit: "px" } },
-    }
-  }
-}